+++ __tests__/interceptors.spec.ts
@@ -1,6 +1,5 @@
-import { defineMiddleware, fatcher } from '../src';
+import { defineMiddleware, fatcher, json } from '../src';
 import fetchMock from 'jest-fetch-mock';
-import { json } from './utils';
 
 const BASE_URL = 'https://fatcher.virtual';
