+++ __tests__/isNodeJS.spec.ts
@@ -1,4 +1,4 @@
-import { isNodeJS } from '../../src/utils/index';
+import { isNodeJS } from '../src/utils';
 
 describe('isNodeJS', () => {
     it('Test env is in node js', () => {
