+++ __tests__/middlewares/aborter/concurrency.spec.ts
@@ -1,6 +1,5 @@
-import { fatcher } from '../../../src';
+import { fatcher, aborter, isAbortError } from '../../../src';
 import fetchMock from 'jest-fetch-mock';
-import { aborter, isAbortError } from '../../../src/middlewares/aborter';
 
 describe('Concurrency', () => {
     const BASE_URL = 'https://fatcher.virual';
