+++ __tests__/scopedRequest.spec.ts
@@ -1,6 +1,5 @@
-import { createScopedRequest, isFatcherError } from '../src';
+import { createScopedRequest, isFatcherError, json } from '../src';
 import fetchMock from 'jest-fetch-mock';
-import { json } from './utils';
 const BASE_URL = 'https://fatcher.virtual';
 
 describe('Scoped Request', () => {
