+++ __tests__/utils.ts
@@ -1,5 +1,3 @@
-import { canActivate, defineMiddleware, Middleware } from '../src';
-
 export function getRandomString(length: number) {
     const STRING_TEMPLATE = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ1234567890';
 
@@ -42,37 +40,4 @@ export function getStringStreamByLength(length: number, chunkSize: number) {
 
 export function sleep(timeout = 500) {
     return new Promise(resolve => setTimeout(resolve, timeout));
-}
-
-
-
-/**
- * A middleware for transforming stream into json
- *
- * If response.body never used, will try to transform into json.
- *
- * But if transform error. Will return a origin response.
- */
-export function json(): Middleware {
-    return defineMiddleware(async (context, next) => {
-        const result = await next();
-
-        if (canActivate(result.data)) {
-            const clonedResponse = result.data.clone();
-
-            try {
-                const data = await clonedResponse.json();
-
-                return Object.assign(result, { data });
-            } catch {
-                /**
-                 * If transform error.
-                 *
-                 * Return origin result.
-                 */
-            }
-        }
-
-        return result;
-    }, 'fatcher-middleware-json');
 }
