+++ src/utils/index.ts
@@ -1,9 +1,7 @@
 export * from './combine';
-export * from './getResponseStatus';
 export * from './immutable';
 export * from './isFunction';
 export * from './isNodeJS';
 export * from './isPlainObject';
 export * from './merge';
 export * from './parseURL';
-export * from './Querystring';
